Task-5. Refactoring
