+++ my-app/src/components/atoms/_index.scss
@@ -1,5 +1,5 @@
 @import './button/button';
-@import './input-text/inputText';
+@import './input-text/input-text';
 @import './link/link';
 @import './header/header';
 @import './divider/divider';
