+++ my-app/src/components/atoms/input-text/_input-text.scss
